lv2: some cleanup
Removes the include of system_config.h from lv2 headers.
Moves lv2_obj to lv2.h
