+++ rpcs3/Emu/Cell/Modules/cellAdec.cpp
@@ -2,7 +2,7 @@
 #include "Emu/System.h"
 #include "Emu/IdManager.h"
 #include "Emu/Cell/PPUModule.h"
-#include "Emu/Cell/lv2/sys_sync.h"
+#include "Emu/Cell/lv2/lv2.h"
 
 #ifdef _MSC_VER
 #pragma warning(push, 0)
