+++ rpcs3/Emu/Cell/Modules/cellGame.cpp
@@ -7,7 +7,7 @@
 #include "Emu/Cell/ErrorCodes.h"
 #include "Emu/Cell/PPUModule.h"
 #include "Emu/Cell/lv2/sys_fs.h"
-#include "Emu/Cell/lv2/sys_sync.h"
+#include "Emu/Cell/lv2/lv2.h"
 
 #include "cellSysutil.h"
 #include "cellMsgDialog.h"
