+++ rpcs3/Emu/Cell/Modules/libmixer.cpp
@@ -1,6 +1,6 @@
 #include "stdafx.h"
 #include "Emu/Cell/PPUModule.h"
-#include "Emu/Cell/lv2/sys_sync.h"
+#include "Emu/Cell/lv2/lv2.h"
 
 #include "cellAudio.h"
 #include "libmixer.h"
