+++ rpcs3/Emu/Cell/lv2/sys_cond.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "sys_sync.h"
+#include "lv2.h"
 #include "sys_mutex.h"
 
 struct lv2_mutex;
