+++ rpcs3/Emu/Cell/lv2/sys_fs.cpp
@@ -1,6 +1,6 @@
 #include "stdafx.h"
-#include "sys_sync.h"
 #include "sys_fs.h"
+#include "lv2.h"
 
 #include "Emu/Cell/PPUModule.h"
 #include "Emu/Cell/PPUThread.h"
