+++ rpcs3/Emu/Cell/lv2/sys_overlay.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "Emu/Cell/PPUAnalyser.h"
-#include "sys_sync.h"
+#include "lv2.h"
 
 struct lv2_overlay final : lv2_obj, ppu_module
 {
