+++ rpcs3/Emu/Cell/lv2/sys_timer.cpp
@@ -2,6 +2,7 @@
 #include "sys_timer.h"
 
 #include "Emu/IdManager.h"
+#include "Emu/system_config.h"
 
 #include "Emu/Cell/ErrorCodes.h"
 #include "Emu/Cell/PPUThread.h"
