+++ rpcs3/Emu/Cell/lv2/sys_uart.cpp
@@ -1,10 +1,11 @@
 #include "stdafx.h"
 
 #include "Emu/System.h"
+#include "Emu/system_config.h"
 #include "Emu/Cell/PPUThread.h"
 #include "Emu/Cell/ErrorCodes.h"
-#include "Emu/Cell/lv2/sys_sync.h"
 #include "Emu/Cell/lv2/sys_rsxaudio.h"
+#include "Emu/Cell/lv2/lv2.h"
 
 #include "sys_uart.h"
